requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
requests==2.31.0
newsapi-python==0.2.7
langchain==0.0.267
huggingface-hub==0.16.4
transformers==4.32.1
gunicorn==21.2.0

